requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
Pillow>=8.2.0
pyyaml>=6.0
tqdm>=4.62.0

# Image processing
opencv-python>=4.5.3
scikit-image>=0.18.0
scipy>=1.7.0

# Visualization and evaluation
matplotlib>=3.4.0
tensorboard>=2.6.0

# Optional dependencies
albumentations>=1.0.0  # Additional image augmentations